LPS-92896 Move to proper package
